+++ casbin/ip_parser/parser/equal.cpp
@@ -11,7 +11,7 @@ namespace casbin {
 bool equal(const IPMask& m1, const IPMask& m2) {
     if(m1.size() != m2.size())
         return false;
-    for(int i = 0 ; i < m1.size() ; i++) {
+    for(size_t i = 0 ; i < m1.size() ; i++) {
         if(m1[i] != m2[i] )
             return false;
     }
