+++ casbin/model/model.cpp
@@ -215,9 +215,9 @@ std::vector<std::vector<std::string>> Model::GetFilteredPolicy(const std::string
     std::vector<std::vector<std::string>> res;
     const std::vector<std::vector<std::string>>& policy = m[sec].assertion_map[p_type]->policy;
     res.reserve(policy.size());
-    for(int i = 0 ; i < policy.size() ; i++){
+    for(size_t i = 0 ; i < policy.size() ; i++){
         bool matched = true;
-        for(int j = 0 ; j < field_values.size() ; j++){
+        for(size_t j = 0 ; j < field_values.size() ; j++){
             if(field_values[j] != "" && (policy[i])[field_index + j] != field_values[j] ){
                 matched = false;
                 break;
@@ -373,9 +373,9 @@ std::pair<bool, std::vector<std::vector<std::string>>> Model::RemoveFilteredPoli
     tmp.reserve(policy.size());
     effects.reserve(policy.size());
     bool res = false;
-    for(int i = 0 ; i < policy.size() ; i++) {
+    for(size_t i = 0 ; i < policy.size() ; i++) {
         bool matched = true;
-        for (int j = 0 ; j < field_values.size() ; j++) {
+        for (size_t j = 0 ; j < field_values.size() ; j++) {
             if (field_values[j] != "" && (policy[i])[field_index+j] != field_values[j]) {
                 matched = false;
                 break;
