+++ casbin/persist/file_adapter/file_adapter.cpp
@@ -41,7 +41,7 @@ void FileAdapter :: SavePolicy(const std::shared_ptr<Model>& model) {
     std::string tmp;
 
     for (std::unordered_map<std::string, std::shared_ptr<Assertion>> :: iterator it = model->m["p"].assertion_map.begin() ; it != model->m["p"].assertion_map.end() ; it++){
-        for (int i = 0 ; i < it->second->policy.size() ; i++){
+        for (size_t i = 0 ; i < it->second->policy.size() ; i++){
             tmp += it->first + ", ";
             tmp += ArrayToString(it->second->policy[i]);
             tmp += "\n";
@@ -49,7 +49,7 @@ void FileAdapter :: SavePolicy(const std::shared_ptr<Model>& model) {
     }
 
     for (std::unordered_map <std::string, std::shared_ptr<Assertion>> :: iterator it = model->m["g"].assertion_map.begin() ; it != model->m["g"].assertion_map.end() ; it++){
-        for (int i = 0 ; i < it->second->policy.size() ; i++){
+        for (size_t i = 0 ; i < it->second->policy.size() ; i++){
             tmp += it->first + ", ";
             tmp += ArrayToString(it->second->policy[i]);
             tmp += "\n";
